Reformat code, Move org.apache.commons.lang3.StringUtils:indexOfAny fork to CsvMatcher (apache2 license as well)
