+++ app/src/main/java/net/gsantner/markor/format/csv/CsvTextConverter.java
@@ -17,13 +17,12 @@ import other.de.stanetz.jpencconverter.JavaPasswordbasedCryption;
 
 /**
  * Part of Markor-Architecture implementing Preview/Export for csv.
- *
+ * <p>
  * Converts csv to md and let
  * {@link  MarkdownTextConverter#convertMarkup(String, Context, boolean, File)}
  * do the rest.
  * <p>
  * This way csv columns may contain md expressions like bold text.
- *
  */
 @SuppressWarnings("WeakerAccess")
 public class CsvTextConverter extends MarkdownTextConverter {
